// src/cif/interface/corpcust.interface.ts
export interface CorpCust {

    street: string
    addCategory?: string
    cityCode: string
    countryCode: string
    stateCode: string
    zipCode: string
    corpName: string
    phoneCityCode: number
    phoneCountryCode: number
    phoneLocalCode: number
    email: string
    createdBy: string
    custHealthCode: string
    incorpDate: string
    keyContactPerson: string
    healthDesc: string
    legalEntityType: string
    entityClass: string
    aveAnnualIncome: number
    primServiceCenter: string
    primRmId: string
    relCreatedBy: string
    relStartDate: string
    shortName: string
    regNumber: string
    taxId: string
    principlePlaceOperation: string
    docCountryOfIssue: string
    docIssDate: string
    docPlaceofIssueCode: string
    docRefNumber: string
    custCrncyCode: string
}